this completely failed
I tried to use reagent to handle the updates.

but I could not get it to work.

Not sure why, but madness madness.

Okay.  quitting because spaghetti, but really I learned a lot here. May
come back to it, but as usual, I did not save the case where the letters
were animating in and out using d3 properly.  That was cool, but it was
eating up lots of CPU.  So I wanted to try to use reagent, but failed.

In a nutshell, instead of creating elements in the d3 loop, I wanted to
render them in reagent,  but i couldn't see how to render them without
resorting to a crazy layer cake of indirections.