spec checking works, except for initial
don't know spec well enough to allow for empty db in initial case, but
spec is working at every db modification...just can't use it for the
shuffle event, because the very first time the db is empty and out of
spec.  As soon as shuffle runs, it is fine I guess.